refactor(nika-event): the journal fold joins the event crate — main's debt paid by the pusher
The push gate caught main itself over two walls (the sister arcs and
my context merge SUMMED): nika-cli at 15068/15000 and settle_ran at
101/100. The pusher repairs:

- nika_event::fold born — RunFact · fold_traces · fold_one move to the
  crate that OWNS the journal format (the workflow_* terminal kinds ·
  the fields envelope). Serde-gated. The CLI keeps the render voice;
  future consumers (MCP · cockpit) fold without the CLI. nika-cli back
  under the cap (14.9k).
- settle_ran folds its retry loop into emit_retry_history — a
  self-contained frame emitter, the 100-line cap forcing the split
  where the block already was one.

Baselines: nika-event +40 lines (fold module · std/serde shapes only —
verified against the ubuntu canonical forms) · nika-cli untouched
(context's structs are module-private).

Gates: crate-size OK · fn-length OK · clippy -D 0 · lib 349+7+105.
